fixed erros with expecting Id type long but receiving type int, tidied up code

+++ src/main/java/com/example/MvRepTile3/app/ExercisesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ExercisesRepository extends JpaRepository<Exercises, Long>{
+public interface ExercisesRepository extends JpaRepository<Exercises, Integer>{
 
 }
